[Web] Fix window size handling
Passing nulls as the first parameters for emscripten_set_resize_callback()
and emscripten_get_element_css_size() was resulting in JavaScript-side
errors due to querySelector() was being passed an empty selector.

This gets around it by passing "#cavas" to the former and a magical
constant which refers to the Window for the latter.